-- added two visual layers:
a. Glitch - Based on ofxPostGlitch (https://github.com/maxillacult/ofxPostGlitch)

this add-on provides at least 7 different effects to be used, we still need to code the GUI for them.

b. GlitchAlt - Based on ofxJpegGlitch (https://github.com/2bbb/ofxJpegGlitch)

this add-on performs basic jpeg glitch (it is a bit nicer that the other), and allows to control a few glitch parameter (GUI not added yet).

-- THIS IS NOT TESTED YET